+++ CHANGELOG.md
@@ -13,7 +13,11 @@ All notable changes to this project will be documented in this file.  The format
 
 ## [Unreleased]
 
+### Add
+* Added a new subcommand `get-era-summary` which optionally takes a block identifier and returns an era summary from a Casper network.
 
+### Changed
+* Update dependencies.
 
 ## [1.5.1] - 2023-03-08